+++ CHANGELOG.md
@@ -6,7 +6,11 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
--   Parser.fromString now also accepts a single Options object if the same options apply to all dice
+-   `Parser.fromString` now also accepts a single Options object if the same options apply to all dice
+
+### Changed
+
+-   Removed option to pass `undefined` to `Parser` for diceThrower instance
 
 ## [0.3.0] - 2021-07-24